+++ src/main/java/cn/nit/beauty/service/PersonService.java
@@ -4,7 +4,7 @@ import cn.nit.beauty.domain.Person;
 
 public interface PersonService {
 	public Person getRandom();
-	public Person getById(Long id);
+	public Person getById(Integer id);
 	public Person save(Person person);
     public Person getByUsername(String username);
